backend/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.3
cycler==0.11.0
distlib==0.3.4
filelock==3.6.0
Flask==2.0.2
fonttools==4.28.5
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
parso==0.8.3
Pillow==8.4.0
platformdirs==2.5.1
protobuf==3.19.4
pyparsing==3.0.6
PyQt3D==5.15.5
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.5
python-dateutil==2.8.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
TBB==0.2
torch==1.10.2+cpu
torchvision==0.11.3+cpu
typing_extensions==4.1.1
urllib3==1.26.6
virtualenv==20.13.2
Werkzeug==2.0.2
yawsso==0.7.1
